add send_non_lagging method for broadcast channel
This method permits users to avoid Lagged broadcast channels.
Historically, there is no way to check if broadcast channels are about
to become Lagged. This prevents senders from being able to easily wait
for all receivers to be ready for a new message, which is problematic
in applications that need to process all messages and can afford to slow
down the sender in doing so.

A new method+error type is introduced to maintain backwards
compatibility and to preserve the more common simple case in which
discarding old messages is acceptable.